Remove op_common.py
op_common.py was used as utils-file, but named as an op.
Functions are moved tosa_utils.py.
